EIP-7702: added AUTH_LIST_LENGTH and new edge cases tests
