+++ engine-tests/src/tests/mod.rs
@@ -27,6 +27,7 @@ mod silo;
 mod standalone;
 mod standard_precompiles;
 mod transaction;
+mod trqnsaction_eip7702;
 pub mod uniswap;
 mod upgrade;
 pub mod xcc;
